Improve purge and export dialogs
There is now a new button to select scenarios and alternatives in
the different purge database dialogs as well as in the export database
dialog. The dialog also shows a warning when any "dangerous" items
have been selected.

The different dialogs have been merged into one to share as much code
as makes sense.

Re #1905