Fix connection related undo/redo commands
Some undo commands related to adding/updating connection and their settings
were storing a reference to the connection itself. However, the connection
becomes obsolete the moment it is removed, because its resource filter model
gets deleted. This means that we cannot store the connections themselves
but all instead all information that is required to find the connections
and restore them as needed.

Re #1976