Fix/rework current database handling in Pivot table
Pivot table was broken in a few ways when database was not frozen.
Trying to edit a database header row or column would raise a Traceback
due to missing `TopLeftDatabaseHeaderItem.add_data()`. Such rows and
column are now uneditable except when the row/column is the "empty" one.
For those rows, user can select into which database the new data is added.
Previously, we used to add data to all available databases.

Re #2064