Fix select new file in Tool Spec Editor
If the spec editor of a tool that has no specification is opened, the text editor widget is completely empty and the header is changed into "No file selected". When a main program file is created, it is automatically selected and the text editor is ready for use. +fixed a bug that sometimes when creating a new main program file, the file would be created, but it would not appear in the spec editor.

Re spine-tools/Spine-Toolbox#2099