Add tooltip warning about long scenario names in db editor
Added a note about scenario names getting shortened down to 20 characters when files with scenario names are generated as a tooltip in the scenario tree. Increased the amount of scenario names' characters in generated subdirectories from 15 to 20. Also removed incorrect tooltip from the last item in the alternative tree.

Re spine-tools/Spine-Toolbox#2081