Fix undo command macro
ManageMembersDialog.accept() built the undo macro that added and removed
entity group members using AgedUndoCommand as the macro's parent.
AgedUndoCommand cannot be used like this, however, since it will not
call the redo() methods of its child commands. This prevented the dialog
that manages entity groups to make any changes to the database.

This commit replaces AgedUndoCommand by SpineDBMacro which is the proper parent
command to use with macros.

Re #2219