Check if item is valid before trying to extract data from it
MultiDBTreeItem may have no db_map_ids for a short interval between the
corresponding db item being removed from cache and the tree item still being
a member of a tree. During this interval it was possible that the UI thread would
try to access the data() mathod of MultiDBTreeItem which would Traceback.

Re #2289,#2294