Fix error while opening project allows to overwrite project
Now if the user tries to open a project that has a version
number that is too low for the Toolbox version, the project
will automatically close, so that the project can't be messed
with. After the project is closed a message is emitted to the
Event Log stating that the project couldn't be opened due to
some error.

Re #2279