Keep track of incremental fetching in the SpineDBWorker
This is because DatabaseMapping now supports filtering in SQL.
We don't use that feature in toolbox, we keep fetching tables
without any filtering; that's generally faster as the same data
is usually needed from several different places. If all those places
run their own filtered queries, then there is a lot of double work.

But, the above also means that DatabaseMapping cannot support
keeping track of the incremental fetching by itself (because it
prefers to support filtered queries that bring data into memory
in a more disorganized way). So, we need to keep track of the offsets
in DBWorker.