Remove before update before add in DB editor
To try and avoid ugly transient states. For example, if I remove and
add stuff and the addition is processed first, some views might
complain they don't find the items I removed. So it's better to process
the removal first so the view is clean when the addition is processed.