Fix copying and pasting boolean values and display_icon
CopyPasteTableView assumed all data in the table was either text or numbers.
This is not true for display_icon (integer) or the active/active_by_default
flags (boolean). We now allow specifying the data types of table columns
in CopyPasteTableView so it has some context to convert the data types properly.

Re spine-tools/Spine-Database-API#316