Fix cx_Freeze bundles, add support for PyInstaller
cx_Freeze and PyInstaller can now be used to bundle Toolbox.

Major change to previous bundling is that we add an official embeddable
Python to the bundle instead of copying the files from the Python that
was used to create the bundle. The previous approach produced an interpreter
that did not always work correctly.