Fix pivot view not loading after URL has changed
The current database gets closed when a new URL is typed to the URL bar
in Database editor. This in turn sets the fetch parents of pivot models
obsolete. We now set the models back to non-obsolete to allow the pivot
table to continue working with the new URL.

This also speeds things up slightly by temporarily disconnecting signals
in a few strategic places which prevents multiple calls to certain functions.

Removed some dead code, too.

Re #2552