Fix removing data from Frozen table
The selected row in Frozen table will change automatically when we
start removing rows from the table because QTableView updates the
current index such that it does not become invalid during the removing.
This could mess up the selected index update mechanism in
FrozenTableModel.remove_values() in certain situations leading to a Traceback.
The method does not depend on the original row selection anymore fixing
the issue.

Re #2552