Apply compatibility transforms in DB manager
Previously, commit_session() would apply compatibility transformations
in spinedb_api and we would then re-apply the changes in DB manager
to introduce them to DB editor and other clients. This approach did
not work very well, however, as reapplying the transforms in DB
manager meant we would add, update and delete same items twice.

Now that spinedb_api does not force the transformations anymore,
we can apply them in DB manager side which works nicely with
clients and the undo stack. Users of DB editor can review the changes
and commit if they are happy or rollback/undo if not.

Re #2625