Disconnect pivot model signals when changing the model
We don't want to connect the same signals multiple times. Also,
if not disconnected, different pivot models may try to add/remove
data from the frozen table which leads to Traceback e.g. when
a pivot model tries to add a row that contains data
to frozen table that doesn't have that column.

Re #2574