Apply changes only for current databases
It is possible that we fetch data that is no longer relevant e.g. when
user opens another database in DB editor while the previous one is not
yet fully loaded. We now drop data from databases that are not current
anymore in compound models to avoid the situation.

Re #2709