Fix tearing down tree model on DB editor close
This fixes a bug where the fetch parents of Alternative and Scenario
models were not properly set as obsolete on DB editor close.

The parents are set obsolete when the model is destroyed by Qt.
However, the slot that the 'destroyed' signal connects to is not always
called when an object is destroyed, see

https://github.com/spine-tools/Spine-Toolbox/issues/2742

To get around the issue, we now connect the signal to a lambda
function.

Re #2742