Add alternative -column to Add entities dialog
Now in the Add entities dialog, there is a column where an
alternative can be specified. If the alternative is specified,
an entity alternative with active=true will be created automatically.
If an alternative is selected from the Alternative tree, that one
will be the default alternative to be offered in the dialog.

Re #2788