Make pasting into tables faster
The row autocompletion is now disabled when data is pasted into
a table. Pasting large amounts of data took a long time if the
db had a lot of items in it already, because fetching the
candidate entity classes from the db_map got slow. Now instead
it is assumed that the user knows what they are doing and
therefore when pasting data, the entity classes are correctly
defined.

Re #2867