Automatically insert rows when pressing down key on last row
CopyPasteTableView now inserts an extra row to the table
when the down key is pressed if its model is EmptyRowModel.
This is handy e.g. in Add Entities or Add Entity Classes dialogs
where, after giving a name for the current entity row, you just
press the down key to start adding another item.