New size hint system for empty tables
The empty tables in DB editor now try to grow as large as possible
when the corresponding data table is smaller than the parent dock.
This means that the empty row right below the last data row
most of the time. There is still a gap due to Qt's layout heuristic
when the data table has only a few rows, though.