Limit pivot table data so DB editor won't hang for a long period
If there would be more than 100 000 elements in Index or Element
pivot, we stop generating the pivot data and show a warning to
user. This prevents DB editor from getting stuck in the generation
process for extended periods of time.

Also, fix a unit test that fails to clean a temporary directory.