Fix file dialog title for Tool
Tool has a button that opens a file dialog for selecting an output
dir. This used a function in toolbox.helpers that has a hard-coded
dialog title referring to Tool's work dir.

This commit makes the helper function take the dialog title as a
parameterer making it more generally useful.