Drop SpineOpt version requirement, try updating always
It was not possible to update SpineOpt even when an update was
available if SpineOpt satisfied REQUIRED_SPINE_OPT_VERSION.
This didn't really make sense: it should always be possible to
update SpineOpt if a more recent version is available.

If the SpineOpt Install/Update Wizard now detects that SpineOpt
is installed, it offers to update it. After attempted update,
it compares the versions of SpineOpt before and after the attempt
notifying user on the last page whether update was successful
or no update was available.

It doesn't seem to be easy to find out if an update is available
using Pkg in Julia, so it is better to just offer the option to
update.